refactor; made RDF path private
