+++ cognite/neat/rules/models/_rdfpath.py
